now track and plot the loss

+++ code/plots.py
@@ -7,7 +7,7 @@ import os
 #for image loading and manipulation
 import cv2
 
-def add_plot_data(accuracy_test,accuracy_train,epoch,run_dir):
+def add_plot_data(accuracy_test,accuracy_train,loss_test,loss_train,epoch,run_dir):
     #creates the directory if it does not exist
     current_dir = run_dir
     output_dir = current_dir + "/plots"
@@ -16,38 +16,54 @@ def add_plot_data(accuracy_test,accuracy_train,epoch,run_dir):
     epochs_file_path = output_dir + "/" + "epochs.npy"
     accuracy_file_path_test = output_dir + "/" + "accuracy_test.npy"
     accuracy_file_path_train = output_dir + "/" + "accuracy_train.npy"
+    loss_file_path_test = output_dir + "/" + "loss_test.npy"
+    loss_file_path_train = output_dir + "/" + "loss_train.npy"
     #if the files exists we load it first
-    if os.path.isfile(epochs_file_path) and os.path.isfile(accuracy_file_path_test) and os.path.isfile(accuracy_file_path_train):
+    if os.path.isfile(epochs_file_path) and os.path.isfile(accuracy_file_path_test) and os.path.isfile(accuracy_file_path_train) and os.path.isfile(loss_file_path_test) and os.path.isfile(loss_file_path_train):
         #loads the existing files
         epochs_numpy_file = np.load(epochs_file_path)
         accuracy_numpy_file_test = np.load(accuracy_file_path_test)
         accuracy_numpy_file_train = np.load(accuracy_file_path_train)
+        loss_numpy_file_test = np.load(loss_file_path_test)
+        loss_numpy_file_train = np.load(loss_file_path_train)
         #appends the new data to the array
         accuracy_numpy_file_test = np.append(accuracy_numpy_file_test,accuracy_test)
         accuracy_numpy_file_train = np.append(accuracy_numpy_file_train,accuracy_train)
+        loss_numpy_file_test = np.append(loss_numpy_file_test,loss_test)
+        loss_numpy_file_train = np.append(loss_numpy_file_train,loss_train)
         epochs_numpy_file = np.append(epochs_numpy_file,epoch)
         #saves the arrays
         np.save(epochs_file_path,epochs_numpy_file)
         np.save(accuracy_file_path_test,accuracy_numpy_file_test)
         np.save(accuracy_file_path_train,accuracy_numpy_file_train)
+        np.save(loss_file_path_test,loss_numpy_file_test)
+        np.save(loss_file_path_train,loss_numpy_file_train)
         print("Added data to numpy file")
     #if the file does not exist we create it
     else:
         epochs_numpy_file = []
         accuracy_numpy_file_test = []
         accuracy_numpy_file_train = []
+        loss_numpy_file_test = []
+        loss_numpy_file_train = []
         #adds the data to the lists
         accuracy_numpy_file_test = np.append(accuracy_numpy_file_test,accuracy_test)
         accuracy_numpy_file_train = np.append(accuracy_numpy_file_train,accuracy_train)
+        loss_numpy_file_test = np.append(loss_numpy_file_test,loss_test)
+        loss_numpy_file_train = np.append(loss_numpy_file_train,loss_train)
         epochs_numpy_file = np.append(epochs_numpy_file,epoch)
         #converts these lists to numpy arrays
         epochs_numpy_file = np.asarray(epochs_numpy_file)
         accuracy_numpy_file_test = np.asarray(accuracy_numpy_file_test)
         accuracy_numpy_file_train = np.asarray(accuracy_numpy_file_train)
+        loss_numpy_file_test = np.asarray(loss_numpy_file_test)
+        loss_numpy_file_train = np.asarray(loss_numpy_file_train)
         #saves the arrays
         np.save(epochs_file_path,epochs_numpy_file)
         np.save(accuracy_file_path_test,accuracy_numpy_file_test)
         np.save(accuracy_file_path_train,accuracy_numpy_file_train)
+        np.save(loss_file_path_test,loss_numpy_file_test)
+        np.save(loss_file_path_train,loss_numpy_file_train)
 
 
 def plot_accuracy(data_directory):
@@ -82,6 +98,35 @@ def plot_accuracy(data_directory):
     plt.xlabel("Epoch")
     plt.ylabel("Accuracy/Error")
     plt.title("Accuracy/Error vs Epoch")
+    plt.legend()
+    plt.show()
+
+def plot_loss(data_directory):
+    print("loss")
+    #used to build error plot
+    error_array_test = []
+    error_array_train = []
+    #builds file paths to load files
+    current_dir = data_directory
+    plot_dir = current_dir + "/" + "plots"
+    epochs_file = plot_dir + "/" + "epochs.npy"
+    accuracy_file_test = plot_dir + "/" + "loss_test.npy"
+    accuracy_file_train = plot_dir + "/" + "loss_train.npy"
+    #loads the arrays
+    epochs_array = np.load(epochs_file)
+    loss_array_test = np.load(accuracy_file_test)
+    loss_array_train = np.load(accuracy_file_train)
+    print(np.amin(loss_array_test))
+    print(np.amin(loss_array_train))
+    #plots and formats accuracy and error for the test data
+    plt.plot(epochs_array,loss_array_test,label="Loss Test")
+    #plt.plot(epochs_array,error_array_test,label="Error Test")
+    #plots and formats the accuracy and error for the train data
+    plt.plot(epochs_array,loss_array_train,label="Loss Train")
+    #plt.plot(epochs_array,error_array_train,label="Error Train")
+    plt.xlabel("Epoch")
+    plt.ylabel("Loss")
+    plt.title("Loss vs Epoch")
     plt.legend()
     plt.show()
